BREAKING: Deprecate monorepo
After a lot of thoughts we decided to manage all UI elements from a
single package as it doesn't makes sense for this project to export all
components as a seperate package on there own.

The previous published package will be deperecated.

- @buildarea-ui/core
- @buildarea-ui/button

Now the new package name would be `@buildarea-ui/react`