fix(Accordion): build fail
I was getting the following error while running build command

```
semantic error TS2686: 'React' refers to a UMD global, but the current
file is a module. Consider adding an import instead.
```

but the Accordion file doesn't have `React.forwardRef` anywhere. So my
understanding was that we don't need that, but dont know still getting
the same error. So after adding seperate default React import solves
this.